requirements.txt
Flask>=3.0
Flask-Cors>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
Werkzeug>=3.0

